Add Issue struct, date sorting
The github issue data structure is very large, so I wanted to trim it
down to something that was as lightweight as it could be with the info
that I wanted to see in the console. Can always extend it later.

The date sorting was an interesting adventure. I think that I should try
to use fewer functions that throw errors, and move toward using more
parsing and GenServers. One thing at a time!

Parse, don't validate, right?